feat: add entities
