+++ src/main/java/com/immobylette/api/main/controller/ThirdPartyController.java
@@ -1,17 +1,22 @@
 package com.immobylette.api.main.controller;
 
 import com.immobylette.api.main.dto.ThirdPartyDto;
+import com.immobylette.api.main.service.ThirdPartyService;
+import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-public class AgentController {
 
+@RestController
+@AllArgsConstructor
+public class ThirdPartyController {
 
+    private final ThirdPartyService thirdPartyService;
 
-    @GetMapping("/agents")
+    @GetMapping("/third-parties/agents")
     public List<ThirdPartyDto> getAgents() {
-        return null;
-        }
+        return thirdPartyService.getAgents();
     }
 }
