+++ src/main/java/com/immobylette/api/main/domain/ThirdPartyTypeEnum.java
@@ -1,2 +1,15 @@
-package com.immobylette.api.main.domain;public enum ThirdPartyTypeEnum {
+package com.immobylette.api.main.domain;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum ThirdPartyTypeEnum {
+
+    AGENT("AGENT"),
+    TENANT("TENANT");
+
+    private final String name;
 }
