+++ src/main/java/com/immobylette/api/main/dto/ThirdPartyDto.java
@@ -1,2 +1,17 @@
-package com.immobylette.api.main.dto;public class ThirdPartyDto {
+package com.immobylette.api.main.dto;
+
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.UUID;
+
+@Builder
+@Getter
+public class ThirdPartyDto {
+
+    private UUID id;
+    private String surname;
+    private String name;
+
 }
