+++ src/main/java/com/immobylette/api/main/entity/ElementType.java
@@ -1,16 +1,14 @@
 package com.immobylette.api.main.entity;
 
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 
+import java.util.List;
 import java.util.UUID;
 
 @Entity
-@Table(name = "third_party_types")
-public class InventoryType {
+@Table(name = "element_types")
+public class ElementType {
 
     @Id
     @Column(name = "id")
@@ -18,5 +16,13 @@ public class InventoryType {
 
     @Column(name = "label")
     private String label;
+
+    @ManyToMany
+    @JoinTable(
+            name = "element_attributes_element_types",
+            joinColumns = @JoinColumn(name = "id_element_attribute"),
+            inverseJoinColumns = @JoinColumn(name = "id_element_type")
+    )
+    private List<ElementAttribute> attributes;
 
 }
