+++ src/main/java/com/immobylette/api/main/entity/HeatingType.java
@@ -9,10 +9,11 @@ import jakarta.persistence.Table;
 import java.util.UUID;
 
 @Entity
-@Table(name = "third_party_types")
-public class HeatingTypes {
+@Table(name = "hot_water_types")
+public class HeatingType {
 
     @Id
+    @Column(name = "id")
     private UUID id;
 
     @Column(name = "label")
