+++ src/main/java/com/immobylette/api/main/entity/SignatureInventoryThirdParty.java
@@ -7,42 +7,23 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
 @Entity
-@Table(name = "third_parties")
+@Table(name = "signatures_inventories_third_parties")
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Inventory {
-    @Id
-    @Column(name = "id")
-    private UUID id;
+public class SignatureInventoryThirdParty {
 
-    @Column(name = "inventory_date")
-    private String inventoryDate;
+    @EmbeddedId
+    private SignatureInventoryThirdPartyId id;
 
-    @JoinColumn(name = "fk_inventory_type")
-    @ManyToOne
-    private InventoryType fkInventoryType;
 
-    @JoinColumn(name = "fk_lease")
-    @ManyToOne
-    private Lease fkLease;
-
-    @JoinColumn(name = "fk_third_party")
-    @ManyToOne
-    private ThirdParty agent;
-
-    @ManyToMany
-    @JoinTable(
-            name = "signatures_inventories_third_parties",
-            joinColumns = @JoinColumn(name = "id_inventory"),
-            inverseJoinColumns = @JoinColumn(name = "id_third_party")
-    )
-    private List<ThirdParty> signatures;
-
+    @Column(name = "signature_date")
+    private Date signatureDate;
 
 }
