+++ src/main/java/com/immobylette/api/main/entity/SignatureInventoryThirdPartyId.java
@@ -1,27 +1,26 @@
 package com.immobylette.api.main.entity;
 
 
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 
-import java.util.List;
+import java.io.Serializable;
 import java.util.UUID;
 
-@Entity
-@Table(name = "signatures_inventories_third_parties")
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class SignatureInventoryThirdParty {
-    @Id
+@Embeddable
+public class SignatureInventoryThirdPartyId implements Serializable {
+
     @Column(name = "id")
     private UUID id;
 
+    @ManyToOne
+    @JoinColumn(name = "id_inventory")
+    private Inventory inventory;
 
-
+    @ManyToOne
+    @JoinColumn(name = "id_third_party")
+    private ThirdParty thirdParty;
 
 }
