+++ src/main/java/com/immobylette/api/main/entity/ThirdPartyType.java
@@ -1,2 +1,21 @@
-package com.immobylette.api.main.entity;public class ThirdPartyType {
+package com.immobylette.api.main.entity;
+
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
+import java.util.UUID;
+
+@Entity
+@Table(name = "third_party_types")
+public class ThirdPartyType {
+
+    @Id
+    private UUID id;
+
+    @Column(name = "label")
+    private String label;
+
 }
