+++ src/main/java/com/immobylette/api/main/mapper/ThirdPartyMapper.java
@@ -1,2 +1,15 @@
-package com.immobylette.api.main.mapper;public interface ThridPartyMapper {
+package com.immobylette.api.main.mapper;
+
+
+import com.immobylette.api.main.dto.ThirdPartyDto;
+import com.immobylette.api.main.entity.ThirdParty;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface ThirdPartyMapper {
+
+    ThirdPartyDto fromThirdParty(ThirdParty thirdParty);
+
 }
+
+
