+++ src/main/resources/db/init/V1__init.sql
@@ -1,0 +1,1 @@
+CREATE EXTENSION IF NOT EXISTS hstore;
